requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-ipware==7.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
Markdown==3.8
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
python-ipware==3.0.0
sqlparse==0.5.3
tzdata==2025.2
